Inclusão das regras de negócio, validando objeto fornecedor antes da persistência.

+++ src/DevIO.Business/Services/FornecedorService.cs
@@ -18,7 +18,13 @@ public class FornecedorService : BaseService, IFornecedorService
         if (!ExecutarValidacao(new FornecedorValidation(), fornecedor)
             || !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco))
             return;
-        //TODO: Validar se não existe outro fornecedor com o mesmo documento
+
+        if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any())
+        {
+            Notificar("Já existe um fornecedor com o Documento informado.");
+            return;
+        }
+
         await _fornecedorRepository.Adicionar(fornecedor);
     }
 
@@ -26,11 +32,39 @@ public class FornecedorService : BaseService, IFornecedorService
     {
         if (!ExecutarValidacao(new FornecedorValidation(), fornecedor))
             return;
+
+        if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento
+            && f.Id != fornecedor.Id).Result.Any())
+        {
+            Notificar("Já existe um fornecedor com o Documento informado.");
+            return;
+        }
         await _fornecedorRepository.Atualizar(fornecedor);
     }
 
     public async Task Remover(Guid id)
     {
+        var fornecedor = await _fornecedorRepository.ObterFornecedorProdutosEndereco(id);
+
+        if (fornecedor == null)
+        {
+            Notificar("Fornecedoor não existe.");
+            return;
+        }
+
+        if (fornecedor.Produtos.Any())
+        {
+            Notificar("O fornecedor possui produtos cadastrados.");
+            return;
+        }
+
+        var endereco = await _fornecedorRepository.ObterEnderecoPorFornecedor(id);
+
+        if (endereco != null)
+        {
+            await _fornecedorRepository.RemoverEnderecoFornecedor(endereco);
+        }
+
         await _fornecedorRepository.Remover(id);
     }
